+++ packages/lib/themes/light.ts
@@ -11,7 +11,7 @@ const theme: Theme = {
 	oddBackgroundColor: '#eeeeee',
 	color: '#32373F', // For regular text
 	colorError: 'red',
-	colorWarn: 'rgb(228 86 0)',
+	colorWarn: 'rgb(228,86,0)',
 	colorWarnUrl: '#155BDA',
 	colorFaded: '#7C8B9E', // For less important text
 	colorBright: '#000000', // For important text