+++ packages/server/src/middleware/ownerHandler.ts
@@ -3,6 +3,6 @@ import { contextSessionId } from '../utils/requestUtils';
 
 export default async function(ctx: AppContext, next: KoaNext): Promise<void> {
 	const sessionId = contextSessionId(ctx, false);
-	if (sessionId) ctx.owner = await ctx.models.session().sessionUser(sessionId);
+	ctx.joplin.owner = sessionId ? await ctx.joplin.models.session().sessionUser(sessionId) : null;
 	return next();
 }