+++ packages/server/src/middleware/routeHandler.ts
@@ -7,17 +7,17 @@ export default async function(ctx: AppContext) {
 	const requestStartTime = Date.now();
 
 	try {
-		const responseObject = await execRequest(ctx.routes, ctx);
+		const responseObject = await execRequest(ctx.joplin.routes, ctx);
 
 		if (responseObject instanceof Response) {
 			ctx.response = responseObject.response;
 		} else if (isView(responseObject)) {
 			const view = responseObject as View;
 			ctx.response.status = view?.content?.error ? view?.content?.error?.httpCode || 500 : 200;
-			ctx.response.body = await ctx.services.mustache.renderView(view, {
-				notifications: ctx.notifications || [],
-				hasNotifications: !!ctx.notifications && !!ctx.notifications.length,
-				owner: ctx.owner,
+			ctx.response.body = await ctx.joplin.services.mustache.renderView(view, {
+				notifications: ctx.joplin.notifications || [],
+				hasNotifications: !!ctx.joplin.notifications && !!ctx.joplin.notifications.length,
+				owner: ctx.joplin.owner,
 			});
 		} else {
 			ctx.response.status = 200;
@@ -25,9 +25,9 @@ export default async function(ctx: AppContext) {
 		}
 	} catch (error) {
 		if (error.httpCode >= 400 && error.httpCode < 500) {
-			ctx.appLogger().error(`${error.httpCode}: ` + `${ctx.request.method} ${ctx.path}` + ` : ${error.message}`);
+			ctx.joplin.appLogger().error(`${error.httpCode}: ` + `${ctx.request.method} ${ctx.path}` + ` : ${error.message}`);
 		} else {
-			ctx.appLogger().error(error);
+			ctx.joplin.appLogger().error(error);
 		}
 
 		// Uncomment this when getting HTML blobs as errors while running tests.
@@ -47,15 +47,15 @@ export default async function(ctx: AppContext) {
 				content: {
 					error,
 					stack: config().showErrorStackTraces ? error.stack : '',
-					owner: ctx.owner,
+					owner: ctx.joplin.owner,
 				},
 				title: 'Error',
 			};
-			ctx.response.body = await ctx.services.mustache.renderView(view);
+			ctx.response.body = await ctx.joplin.services.mustache.renderView(view);
 		} else { // JSON
 			ctx.response.set('Content-Type', 'application/json');
 			const r: any = { error: error.message };
-			if (ctx.env === Env.Dev && error.stack) r.stack = error.stack;
+			if (ctx.joplin.env === Env.Dev && error.stack) r.stack = error.stack;
 			if (error.code) r.code = error.code;
 			ctx.response.body = r;
 		}
@@ -63,6 +63,6 @@ export default async function(ctx: AppContext) {
 		// Technically this is not the total request duration because there are
 		// other middlewares but that should give a good approximation
 		const requestDuration = Date.now() - requestStartTime;
-		ctx.appLogger().info(`${ctx.request.method} ${ctx.path} (${requestDuration}ms)`);
+		ctx.joplin.appLogger().info(`${ctx.request.method} ${ctx.path} (${requestDuration}ms)`);
 	}
 }