+++ packages/server/src/models/ShareUserModel.ts
@@ -1,6 +1,7 @@
 import { Item, Share, ShareType, ShareUser, ShareUserStatus, User, Uuid } from '../db';
 import { ErrorForbidden, ErrorNotFound } from '../utils/errors';
 import BaseModel, { AclAction, DeleteOptions } from './BaseModel';
+import { getCanShareFolder } from './utils/user';
 
 export default class ShareUserModel extends BaseModel<ShareUser> {
 
@@ -10,8 +11,8 @@ export default class ShareUserModel extends BaseModel<ShareUser> {
 
 	public async checkIfAllowed(user: User, action: AclAction, resource: ShareUser = null): Promise<void> {
 		if (action === AclAction.Create) {
-			const recipient = await this.models().user().load(resource.user_id, { fields: ['can_share_folder'] });
-			if (!recipient.can_share_folder) throw new ErrorForbidden('The sharing feature is not enabled for the recipient account');
+			const recipient = await this.models().user().load(resource.user_id, { fields: ['account_type', 'can_share_folder'] });
+			if (!getCanShareFolder(recipient)) throw new ErrorForbidden('The sharing feature is not enabled for the recipient account');
 
 			const share = await this.models().share().load(resource.share_id);
 			if (share.owner_id !== user.id) throw new ErrorForbidden('no access to the share object');