+++ packages/server/src/routes/index/shares.ts
@@ -24,19 +24,19 @@ const router: Router = new Router(RouteType.Web);
 router.public = true;
 
 router.get('shares/:id', async (path: SubPath, ctx: AppContext) => {
-	const shareModel = ctx.models.share();
+	const shareModel = ctx.joplin.models.share();
 
 	const share = await shareModel.load(path.id);
 	if (!share) throw new ErrorNotFound();
 
-	const itemModel = ctx.models.item();
+	const itemModel = ctx.joplin.models.item();
 
 	const item = await itemModel.loadWithContent(share.item_id);
 	if (!item) throw new ErrorNotFound();
 
 	const result = await renderItem(ctx, item, share);
 
-	ctx.models.share().checkShareUrl(share, ctx.URL.origin);
+	ctx.joplin.models.share().checkShareUrl(share, ctx.URL.origin);
 
 	ctx.response.body = result.body;
 	ctx.response.set('Content-Type', result.mime);