+++ packages/server/src/routes/index/signup.test.ts
@@ -1,6 +1,7 @@
 import config from '../../config';
 import { NotificationKey } from '../../models/NotificationModel';
 import { AccountType } from '../../models/UserModel';
+import { getCanShareFolder, getMaxItemSize } from '../../models/utils/user';
 import { MB } from '../../utils/bytes';
 import { execRequestC } from '../../utils/testing/apiUtils';
 import { beforeAllDb, afterAllTests, beforeEachDb, models } from '../../utils/testing/testUtils';
@@ -43,8 +44,8 @@ describe('index_signup', function() {
 		expect(user).toBeTruthy();
 		expect(user.account_type).toBe(AccountType.Basic);
 		expect(user.email_confirmed).toBe(0);
-		expect(user.can_share_folder).toBe(0);
-		expect(user.max_item_size).toBe(10 * MB);
+		expect(getCanShareFolder(user)).toBe(0);
+		expect(getMaxItemSize(user)).toBe(10 * MB);
 
 		// Check that the user is logged in
 		const session = await models().session().load(context.cookies.get('sessionId'));