+++ packages/server/src/routes/index/signup.ts
@@ -8,7 +8,7 @@ import defaultView from '../../utils/defaultView';
 import { View } from '../../services/MustacheService';
 import { checkPassword } from './users';
 import { NotificationKey } from '../../models/NotificationModel';
-import { AccountType, accountTypeProperties } from '../../models/UserModel';
+import { AccountType } from '../../models/UserModel';
 import { ErrorForbidden } from '../../utils/errors';
 
 function makeView(error: Error = null): View {
@@ -43,17 +43,17 @@ router.post('signup', async (_path: SubPath, ctx: AppContext) => {
 		const formUser = await bodyFields<FormUser>(ctx.req);
 		const password = checkPassword(formUser, true);
 
-		const user = await ctx.models.user().save({
-			...accountTypeProperties(AccountType.Basic),
+		const user = await ctx.joplin.models.user().save({
+			account_type: AccountType.Basic,
 			email: formUser.email,
 			full_name: formUser.full_name,
 			password,
 		});
 
-		const session = await ctx.models.session().createUserSession(user.id);
+		const session = await ctx.joplin.models.session().createUserSession(user.id);
 		ctx.cookies.set('sessionId', session.id);
 
-		await ctx.models.notification().add(user.id, NotificationKey.ConfirmEmail);
+		await ctx.joplin.models.notification().add(user.id, NotificationKey.ConfirmEmail);
 
 		return redirect(ctx, `${config().baseUrl}/home`);
 	} catch (error) {