+++ packages/server/src/utils/types.ts
@@ -2,6 +2,7 @@ import { LoggerWrapper } from '@joplin/lib/Logger';
 import * as Koa from 'koa';
 import { DbConnection, User, Uuid } from '../db';
 import { Models } from '../models/factory';
+import { Account } from '../models/UserModel';
 import { Services } from '../services/types';
 import { Routers } from './routeUtils';
 
@@ -18,15 +19,32 @@ export interface NotificationView {
 	closeUrl: string;
 }
 
-export interface AppContext extends Koa.Context {
+interface AppContextJoplin {
 	env: Env;
 	db: DbConnection;
 	models: Models;
 	appLogger(): LoggerWrapper;
 	notifications: NotificationView[];
 	owner: User;
+	account: Account;
 	routes: Routers;
 	services: Services;
+}
+
+export interface AppContext extends Koa.Context {
+	joplin: AppContextJoplin;
+
+	// All the properties under `joplin` were previously at the root, so to make
+	// sure they are no longer used anywhere we set them to "never", as that
+	// would trigger the TypeScript compiler. Later on, all this can be removed.
+	env: never;
+	db: never;
+	models: never;
+	appLogger: never;
+	notifications: never;
+	owner: never;
+	routes: never;
+	services: never;
 }
 
 export enum DatabaseConfigClient {